// shared/models/user.model.ts
export class User {
  AuthInfo: any;
  operationStatus: string;
  operationMessage: string;
  // status:boolean;
  item: {
    clientMasterId: number;
    clientMasterName: string;
    email: string;
    token: string;
    phone: string;
    userId: number;
    userName: string;
    firstName: string;
    lastName: string;
    userAccountId: string;
    appRoles: string;
    clientlogopath: string;
    userimagepath: string;
    userGroupId: string;
    primaryUser: string;
    rcApp:boolean;
    countryCode:string;
  };
}